refactor(behat): don't use subtree split approach
