+++ .github/workflows/static-analysis.yml
@@ -47,8 +47,10 @@ jobs:
       - name: Install PHPStan
         run: composer bin phpstan install
 
-      - name: Install PHPBench
-        run: composer bin phpbench install
+      - name: Install PHPBench & Behat
+        run: |
+          composer bin phpbench install
+          composer bin behat install
 
       - name: Run PHPStan
         run: bin/tools/phpstan/vendor/phpstan/phpstan/phpstan analyse
