+++ bin/tools/behat/composer.json
@@ -1,0 +1,14 @@
+{
+    "require": {
+        "behat/behat": "^3.22",
+        "behat/mink-browserkit-driver": "^2.0",
+        "friends-of-behat/mink-extension": "^2.0",
+        "friends-of-behat/symfony-extension": "^2.0",
+        "symfony/flex": "^2.10"
+    },
+    "config": {
+        "allow-plugins": {
+            "symfony/flex": true
+        }
+    }
+}
