--- src/Behat/.gitattributes
@@ -1,8 +1,0 @@
-* text=auto
-
-/.gitattributes                 export-ignore
-/.gitignore                     export-ignore
-/behat.yml                      export-ignore
-/features                       export-ignore
-/symfony.lock                   export-ignore
-/tests                          export-ignore
