--- src/Behat/.gitignore
@@ -1,4 +1,0 @@
-/composer.lock
-/vendor/
-/var/
-/.env.local
