+++ tests/Fixture/Behat/BehatTestKernel.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Zenstruck\Foundry\Behat\Tests;
+namespace Zenstruck\Foundry\Tests\Fixture\Behat;
 
 use FriendsOfBehat\SymfonyExtension\Bundle\FriendsOfBehatSymfonyExtensionBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
-use Zenstruck\Foundry\Behat\Tests\App\HelloWorldController;
+use Zenstruck\Foundry\Tests\Fixture\App\Controller\HelloWorldController;
 use Zenstruck\Foundry\Tests\Fixture\FoundryTestKernel;
 
 final class BehatTestKernel extends FoundryTestKernel
@@ -23,16 +23,6 @@ final class BehatTestKernel extends FoundryTestKernel
         parent::configureContainer($c, $loader);
 
         $c->register(HelloWorldController::class)->setAutowired(true)->setAutoconfigured(true)->addTag('controller.service_arguments');
-        $c->register(FeatureContext::class)->setAutowired(true)->setAutoconfigured(true);
-    }
-
-    protected function configureRoutes(RoutingConfigurator $routes): void
-    {
-        $routes->import(__DIR__.'/App/*.php', 'attribute');
-    }
-
-    protected function baseFixturePath(): string
-    {
-        return '%kernel.project_dir%/../../tests/Fixture';
+        $c->register(TestContext::class)->setAutowired(true)->setAutoconfigured(true);
     }
 }
